Adjust how package data is accessed
Instead of creating a sub-package just to act as a file container we
instead adjust the resource being accessed. This will prevent namespace
cluttering while still providing access to the package data that is
needed. This also follows logically with how the filesystem is
organized where `data/secret.txt` is a resource of the `example`
package.